requirements.txt
fastapi
uvicorn
requests
python-dotenv


# fastapi
# uvicorn
# python-telegram-bot
# python-dotenv
# requests